requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128

flask~=3.1.1
python-dotenv~=1.0.0
mistral_common~=1.8.1
transformers~=4.53.2
torch~=2.7.1
huggingface_hub~=0.33.4
accelerate~=1.9.0
numpy~=1.26.4
openai~=1.90.0